UML v1.0
contains first version of the ethereum and Tezos UML Diagrams. Ethereum being PoW requires a back and forth 'handshake' to avoid simultaneous buyers problem. Tezos being PoS in theory does not have to deal with this issue. In reality, Tezos only supports 40 transactions per second compared to Ethereum's 15 so it is still would be a problem. Given that we can't use Cardano yet (1000+ transactions per second), We are using Tezos as an example for what PoS can bring to smart contracts.